colblk: widen Decode function offset/count types to prevent overflow
The Decode functions in the colblk package (DecodeBitmap, DecodeRawBytes,
DecodePrefixBytes, DecodeUnsafeUints, DecodeUnsafeOffsets) previously used
uint32 for offsets and int for counts. Arithmetic on these values (e.g.
off + uint32(rows*w)) was prone to overflow.

Change the offset parameter and return type from uint32 to uint64, and the
count/rows parameter from int to uint32. This eliminates overflow in offset
arithmetic while keeping counts bounded to the uint32 range which is
sufficient for row counts. Note that offsets are checked against the
length of the buffer so they can't get large enough to cause 64-bit
overflows.

Also add bounds checks: an overflow check for count+nBundles in
DecodePrefixBytes, and an endOffset vs len(b) check in DecodeRawBytes.

The DecodeFunc type, DecodeColumn, and pageStart are updated accordingly.
All callers are updated with appropriate casts.

+++ sstable/colblk/bitmap_test.go
@@ -55,7 +55,7 @@ func TestBitmapFixed(t *testing.T) {
 			if endOffset != size {
 				td.Fatalf(t, "endOffset=%d size=%d", endOffset, size)
 			}
-			bitmap, _ = DecodeBitmap(data, off, n)
+			bitmap, _ = DecodeBitmap(data, uint64(off), uint32(n))
 			dumpBitmap(&buf, bitmap)
 			fmt.Fprint(&buf, "\nBinary representation:\n")
 			f := binfmt.New(data)
@@ -174,8 +174,8 @@ func TestBitmapRandom(t *testing.T) {
 
 		data := make([]byte, builder.Size(size, 0))
 		_ = builder.Finish(0, size, 0, data)
-		bitmap, endOffset := DecodeBitmap(data, 0, size)
-		require.Equal(t, uint32(len(data)), endOffset)
+		bitmap, endOffset := DecodeBitmap(data, 0, uint32(size))
+		require.Equal(t, uint64(len(data)), endOffset)
 		for i := 0; i < size; i++ {
 			if got := bitmap.At(i); got != v[i] {
 				t.Fatalf("b.Get(%d) = %t; want %t", i, got, v[i])
